+++ src/controllers/AdminDocumentController.java
@@ -11,13 +11,20 @@ import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXComboBox;
 import com.jfoenix.controls.JFXListView;
 import com.jfoenix.controls.JFXTextField;
+import java.util.Optional;
+import java.util.function.UnaryOperator;
 import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.TextFormatter;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.image.Image;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
+import javafx.stage.Stage;
 
 /**
  * FXML Controller class
@@ -56,6 +63,7 @@ public class AdminDocumentController {
          initComboBox();
          initButtonsActions();
          initKeyBoardsActions();
+         initTextFields();
     }    
     
     public void initTableView(){
@@ -116,24 +124,51 @@ public class AdminDocumentController {
         categDoc_listView.setOnKeyPressed((KeyEvent event) -> {
             KeyCode key = event.getCode();
             if(key == KeyCode.DELETE){
-                CategorieDocument cd = categDoc_listView.getSelectionModel().getSelectedItem();
-                if(cd != null){
-                    categDoc_listView.getItems().remove(cd);
-                    cd.delete();
-                        }
-                };           
+                if(displayDeleteConfirmation()){
+                    CategorieDocument cd = categDoc_listView.getSelectionModel().getSelectedItem();
+                    if(cd != null){
+                        categDoc_listView.getItems().remove(cd);
+                        cd.delete();
+                            }
+                    };   
+            }
         });
         typeDoc_tableView.setOnKeyPressed((KeyEvent event) -> {
             KeyCode key = event.getCode();
             if(key == KeyCode.DELETE){
-                TypeDocument td = typeDoc_tableView.getSelectionModel().getSelectedItem();
-                if(td != null){
-                    typeDoc_tableView.getItems().remove(td);
-                    td.delete();
-                        }
-                }; 
+                if(displayDeleteConfirmation()){
+                    TypeDocument td = typeDoc_tableView.getSelectionModel().getSelectedItem();
+                    if(td != null){
+                        typeDoc_tableView.getItems().remove(td);
+                        td.delete();
+                            }
+                    }; 
+                }
         });
-    }  
+    }
+    
+    public void initTextFields(){
+        UnaryOperator<TextFormatter.Change> filter = (TextFormatter.Change t) -> {
+            String newText = t.getControlNewText();
+            if(newText.matches("^([^\\s].*)*$") && newText.length()< 120) {
+                return t ;
+            }
+            return null ;
+        };
+        categDoc_textField.setTextFormatter(new TextFormatter<>(filter));
+        typeDoc_textField.setTextFormatter(new TextFormatter<>(filter));
+    
+    }
+    public boolean displayDeleteConfirmation(){       
+        Alert dialogConfirm = new Alert(Alert.AlertType.CONFIRMATION);
+        ((Stage)dialogConfirm.getDialogPane().getScene().getWindow()).getIcons().add(new Image("/ressources/images/icon_lawyer2.png"));
+        dialogConfirm.setTitle("Confirmation de suppression");
+        dialogConfirm.setHeaderText("Confirmation de suppression");
+        dialogConfirm.setContentText("Voulez vous vraiment supprimer cet élément ??");
+        Optional<ButtonType> answer = dialogConfirm.showAndWait();
+      
+        return (answer.get() == ButtonType.OK) ;         
+    }
     
 }
